--- sdformat_mjcf/src/sdformat_mjcf/__init__.py
@@ -1,1 +1,0 @@
-# blank
